fix: Bug with get_ingredients route

+++ backend/src/models.rs
@@ -78,7 +78,7 @@ pub struct Ingredient {
     pub category: String,
 
     #[serde(rename = "Quantity")]
-    pub quantity: u32,
+    pub quantity: u16,
 
     #[serde(rename = "Unit")]
     pub unit: String,
